add eval ${val} and contexts
