+++ example/zk.go
@@ -7,7 +7,9 @@ import (
     zkvs "github.com/tietang/props/zk"
 )
 
-func main6() {
+func main() {
+
+
     context := "/configurations/demo/dev/app1"
 
     c, ch, err := zk.Connect([]string{"172.16.1.248:2181"}, 2*time.Second)
@@ -20,17 +22,16 @@ func main6() {
         if event.State == zk.StateConnected {
             break
         }
-
     }
 
-    z := zkvs.NewZookeeperConfigSource("zk", context, c)
+    z := zkvs.NewZookeeperConfigSource("zk", true, context, c)
     v, err := z.Get("app.xx.key-0")
     fmt.Println(v)
     fmt.Println(err)
     z.Watch("app/xx", func(children []string, event zk.Event) {
         fmt.Println("Watch:   ", event, len(z.Keys()))
         for _, key := range z.Keys() {
-            fmt.Println("new value:  ", z.GetDefault(key, key))
+            fmt.Println("new value:  ", key, "= ", z.GetDefault(key, key))
 
         }
     })
