+++ zk/zk_props_source.go
@@ -3,11 +3,8 @@ package zk
 import (
     "github.com/samuel/go-zookeeper/zk"
     "strings"
-    "errors"
     "path"
-    "fmt"
     log "github.com/sirupsen/logrus"
-    "bytes"
     "github.com/tietang/props/kvs"
 )
 
@@ -37,29 +34,24 @@ zk nodes:
 
 
 */
+
 type ZookeeperPropsConfigSource struct {
-    kvs.MapProperties
-    name    string
-    conn    *zk.Conn
-    context string
+    ZookeeperSource
 }
 
-func NewZookeeperPropsConfigSource(name string, context string, conn *zk.Conn) *ZookeeperPropsConfigSource {
+func NewZookeeperPropsConfigSource(name string, watched bool, context string, conn *zk.Conn) *ZookeeperPropsConfigSource {
     s := &ZookeeperPropsConfigSource{}
+    s.Watched = watched
     s.name = name
     s.Values = make(map[string]string)
     s.conn = conn
     s.context = context
-    s.init()
+    s.initProperties()
     return s
 }
 
-func (s *ZookeeperPropsConfigSource) init() {
+func (s *ZookeeperPropsConfigSource) initProperties() {
     s.findProperties(s.context)
-}
-
-func (s *ZookeeperPropsConfigSource) Close() {
-    s.conn.Close()
 }
 
 func (s *ZookeeperPropsConfigSource) findProperties(root string) {
@@ -71,10 +63,14 @@ func (s *ZookeeperPropsConfigSource) findProperties(root string) {
 
         fp := path.Join(root, p)
         value, err := s.getPropertiesValue(fp)
+        if s.Watched && strings.HasSuffix(fp, DEFAULT_WATCH_KEY) {
+            log.Debug("WatchNodeDataChange: ", fp)
+            s.watchGet(fp)
+        }
 
         if err == nil {
             props := kvs.NewProperties()
-            props.Load(bytes.NewReader(value))
+            props.Load(strings.NewReader(value))
             for _, key := range props.Keys() {
                 val := props.GetDefault(key, "")
                 pkey := strings.Join([]string{p, key}, ".")
@@ -84,84 +80,4 @@ func (s *ZookeeperPropsConfigSource) findProperties(root string) {
 
     }
 
-}
-
-func (s *ZookeeperPropsConfigSource) getPropertiesValue(path string) ([]byte, error) {
-    d, _, err := s.conn.Get(path)
-    if err != nil || len(d) == 0 {
-        return nil, errors.New("not value")
-    }
-    return d, nil
-}
-
-func (s *ZookeeperPropsConfigSource) getChildren(childPath string) []string {
-    children, _, err := s.conn.Children(childPath)
-    if err != nil {
-        return make([]string, 0)
-    }
-    return children
-}
-
-func (s *ZookeeperPropsConfigSource) sanitizeKey(path string, context string) string {
-    key := strings.Replace(path, context+"/", "", -1)
-    key = strings.Replace(key, "/", ".", -1)
-    return key
-}
-
-func (s *ZookeeperPropsConfigSource) registerKeyValue(path, value string) {
-    key := s.sanitizeKey(path, s.context)
-    s.Set(key, value)
-
-}
-
-func (s *ZookeeperPropsConfigSource) Name() string {
-    return s.name
-}
-
-func (s *ZookeeperPropsConfigSource) Watch(key string, handlers ... func([]string, zk.Event)) {
-    go s.watchGet(path.Join(s.context, key, KEY_NOTIFY_NODE), handlers...)
-}
-
-func (s *ZookeeperPropsConfigSource) WatchChildren(key string, handlers ... func([]string, zk.Event)) {
-    pathStr := path.Join(s.context, key)
-    s.watchChildren(pathStr, handlers...)
-}
-
-func (s *ZookeeperPropsConfigSource) watchChildren(pathStr string, handlers ... func([]string, zk.Event)) {
-    children, stat, ch, err := s.conn.ChildrenW(pathStr)
-    if err != nil {
-        panic(err)
-    }
-    fmt.Printf("%+v %+v\n", children, stat)
-    e := <-ch
-
-    s.findProperties(e.Path)
-    for _, handler := range handlers {
-        handler(children, e)
-    }
-    fmt.Printf("%+v\n", e)
-    s.watchChildren(pathStr, handlers...)
-}
-
-func (g *ZookeeperPropsConfigSource) watchGet(pathStr string, handlers ... func([]string, zk.Event)) {
-    log.Info(pathStr)
-    exists, _, _ := g.conn.Exists(pathStr)
-    if !exists {
-        g.conn.Create(pathStr, []byte("1"), 1, nil)
-    }
-    _, stat, ch, err := g.conn.GetW(pathStr)
-    children, _, err := g.conn.Children(pathStr)
-    if err != nil {
-        panic(err)
-    }
-    log.Infof("watch: %+v %+v\n", children, stat)
-    e := <-ch
-
-    //pPath:=path.Dir(e.Path)
-    g.findProperties(e.Path)
-    for _, handler := range handlers {
-        handler(children, e)
-    }
-    log.Infof("notify event: %+v\n ", e)
-    g.watchGet(pathStr, handlers...)
 }
